+++ packages/common/components/SortPopoverMenu/fragments/SortPopoverMenuItem.tsx
@@ -36,8 +36,15 @@ const SortPopoverMenuItem = ({ i18nKey, selected, onSelect, value, ...props }: S
     onSelect(value);
   };
 
+  const preventPropagationOnSpaceKey = (event: KeyboardEvent) => {
+    // Prevent conflict between selecting sorting option and push to talk.
+    if (event.code === "Space") {
+      event.stopPropagation();
+    }
+  }
+
   return (
-    <ContainerMenuItem {...props} onClick={onClickHandler}>
+    <ContainerMenuItem {...props} onClick={onClickHandler} onKeyUp={preventPropagationOnSpaceKey} onKeyDown={preventPropagationOnSpaceKey}>
       <Typography>{t(i18nKey)}</Typography>
       <DoneIconStyledWrapper fontSize="inherit" isActive={selected}>
         <DoneIcon />